requirements.txt
streamlit
pandas
requests
aiohttp
plotly
google-generativeai
python-dotenv
elevenlabs
